refactor(language): split moniker handling into dedicated module
